+++ packages/radonis-hooks/src/hooks/isomorphic/use_route.ts
@@ -38,7 +38,7 @@ export function useRoute() {
   }
 
   return {
-    current: route,
+    current: route!,
     isCurrent,
   }
 }